requirements.txt
langchain
faiss-cpu
openai
unstructured
tiktoken
